Some fixes for the repacking and neiblist size setting

+++ src/Options.h
@@ -69,7 +69,7 @@ public:
 	double	deltap; ///< deltap
 	float	tend; ///< simulation end
 	unsigned long maxiter; ///< maximum number of iterations to run
-	unsigned long repack_maxiter; ///< maximum number of iterations for repacking
+	uint repack_maxiter; ///< maximum number of iterations for repacking
 	float	checkpoint_freq; ///< frequency of hotstart checkpoints (in simulated seconds)
 	int		checkpoints; ///< number of hotstart checkpoints to keep
 	bool	nosave; ///< disable saving
