+++ src/simparams.h
@@ -215,7 +215,7 @@ typedef struct SimParams {
 	 * \default{200}
 	 * TLT_REPACK_MAX_ITER
 	 */
-	unsigned long			repack_maxiter; //< maximum number of iterations for repacking
+	uint			repack_maxiter; //< maximum number of iterations for repacking
 	/*!
 	 * \inpsection{repacking, enable}
 	 * \label{REPACK_A}
